Check channel type (#97)

+++ commands/config/configjoin.js
@@ -40,7 +40,7 @@ module.exports = class extends Command {
         const channel = confChannel.mentions.channels.first()
         || message.guild.channels.cache.get(confChannel.content)
         || message.guild.channels.cache.find((ch) => ch.name === confChannel.content || `#${ch.name}` === confChannel.content);
-        if (!channel) return msg.error("config/configjoin:CHANNEL_NOT_FOUND", {
+        if (!channel || channel.type === "voice") return msg.error("config/configjoin:CHANNEL_NOT_FOUND", {
             channel: confChannel.content
         }, true);
         collected.first().delete();
